Support @google/genai v2
