+++ library/package.json
@@ -62,7 +62,8 @@
     "@fastify/cookie": "^11.0.2",
     "@google-cloud/functions-framework": "^5.0.0",
     "@google-cloud/pubsub": "^5.0.0",
-    "@google/genai": "1.14",
+    "@google/genai-v1": "npm:@google/genai@1.52.0",
+    "@google/genai-v2": "npm:@google/genai@2.4.0",
     "@graphql-tools/executor": "^1.3.2",
     "@hapi/hapi": "^21.3.10",
     "@hono/node-server": "^1.12.2",
