+++ library/sinks/GoogleGenAi.v1.test.ts
@@ -1,0 +1,3 @@
+import { createGoogleGenAiTests } from "./GoogleGenAi.tests";
+
+createGoogleGenAiTests("@google/genai-v1");
